feat: Enable music toggle in menu
This commit adds the ability to toggle the music on and off in the menu. It introduces a new variable `music_on` in the `variables.lua` file and updates the `update.lua` file to display the appropriate menu item based on the current state of `music_on`. When the music is turned off, the game stops playing the music, and when it is turned on, the game resumes playing the music.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature).