feat: Update gif timer variable name and value
This commit updates the variable name and value for the gif timer in the `index.html` file. The variable name is changed from `flutter-valley-tourney.js` to `flutter-tourney.js` to reflect the correct file name. Additionally, the value of the variable `sessions` in the `config.txt` file is incremented from 113 to 114 to track the number of times the program has been run.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature).